requirements.txt
django-push-notifications
certifi==2017.11.5
chardet==3.0.4
contextlib2==0.5.5
Django==1.11
django-classy-tags==0.8.0
django-cookie-law==1.0.13
django-sekizai==0.10.0
idna==2.5
MySQL-python==1.2.5
Pillow==5.1.0
pytz==2017.3
raven==6.8.0
requests==2.18.1
urllib3==1.21.1
